Merge dev into main

+++ machine-factory/src/main/java/no/pgr209/machinefactory/controller/AddressController.java
@@ -1,8 +1,11 @@
 package no.pgr209.machinefactory.controller;
 
 import no.pgr209.machinefactory.model.Address;
+import no.pgr209.machinefactory.model.AddressDTO;
 import no.pgr209.machinefactory.service.AddressService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,33 +24,66 @@ public class AddressController {
 
     //Get all addresses
     @GetMapping()
-    public List<Address> getAllAddresses() {
-        return addressService.getAllAddresses();
+    public ResponseEntity<List<Address>> getAllAddresses() {
+        List<Address> allAddresses = addressService.getAllAddresses();
+
+        if(!allAddresses.isEmpty()){
+            return new ResponseEntity<>(allAddresses, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     //Get addresses by page
     @GetMapping("/page/{pageNr}")
-    public List<Address> getAddressesByPage(@PathVariable int pageNr) {
-        return addressService.getAddressesByPage(pageNr);
+    public ResponseEntity<List<Address>> getAddressesByPage(@PathVariable int pageNr) {
+        List<Address> addressesByPage = addressService.getAddressesByPage(pageNr);
+
+        if(!addressesByPage.isEmpty()) {
+            return new ResponseEntity<>(addressesByPage, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("/{id}")
-    public Address getAddressById(@PathVariable Long id) {
-        return addressService.getAddressById(id);
+    public ResponseEntity<Address> getAddressById(@PathVariable Long id) {
+        Address addressById = addressService.getAddressById(id);
+
+        if(addressById != null) {
+            return new ResponseEntity<>(addressById, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping
-    public Address createAddress(Address address) {
-        return addressService.createAddress(address);
+    public ResponseEntity<Address> createAddress(@RequestBody AddressDTO addressDTO) {
+        Address createdAddress = addressService.createAddress(addressDTO);
+
+        if(createdAddress != null) {
+            return new ResponseEntity<>(createdAddress, HttpStatus.CREATED);
+        } else {
+            HttpHeaders responseHeaders = new HttpHeaders();
+            responseHeaders.set("Error", "One of more fields are invalid");
+            return new ResponseEntity<>(responseHeaders, HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteAddressById(@PathVariable Long id) {
-        addressService.deleteAddressById(id);
+    public ResponseEntity<String> deleteAddressById(@PathVariable Long id) {
+        if (addressService.addressExists(id)) {
+            addressService.deleteAddressById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Address> updateAddress(@PathVariable Long id, @RequestBody Address updatedAddress) {
-        return addressService.updateAddress(id, updatedAddress);
+    public ResponseEntity<Address> updateAddress(@PathVariable Long id, @RequestBody AddressDTO addressDTO) {
+        Address updatedAddress = addressService.updateAddress(id, addressDTO);
+
+        if(updatedAddress != null && updatedAddress.getAddressCity() != null && updatedAddress.getAddressStreet() != null && updatedAddress.getAddressZip() != null) {
+            return new ResponseEntity<>(updatedAddress, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
